Set InlineSvg to raise an error on missing file
This configuration will cause an exception to be raised if an Inline SVG
is embedded and the file doesn't exist.

This behavioural change is consistent with what would happen were we to
try import a view that didn't exist or embed an asset pipeline link.

This has been added in as part of my walk of shame for accidentally
linking to a missing image.